Move StartIndexHandler to index package
